💄 Improve the feedback visibility
Beforehand, the feedback message appeared at the bottom of the screen,
which made it difficult to notice on smaller screens.

Now, the feedback is displayed at overlay on the flag, ensuring
it's always visible.

Finally, the border of the flag container is colored to easily
know if the flag is correct or not.